break-out VarianceLegend, VerticalBound, and XAxis
